refactor(cfd-2d): Decompose venturi_flow/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
