+++ crates/cfd-2d/src/solvers/venturi_flow/mod.rs
@@ -64,483 +64,12 @@
 //! monotonically. Convergence is guaranteed by the spectral radius of the iteration matrix
 //! being strictly less than 1.
 
+mod analytical;
+mod geometry;
+mod solution;
 mod solver;
 
+pub use analytical::{BernoulliVenturi, ViscousVenturi};
+pub use geometry::VenturiGeometry;
+pub use solution::VenturiFlowSolution;
 pub use solver::{VenturiSolver2D, VenturiValidationResult, VenturiValidator};
-
-use crate::scalar::Cfd2dScalar;
-use crate::scalar::{self, from_f64};
-use eunomia::{FloatElement, NumericElement};
-use serde::{Deserialize, Serialize};
-
-// ============================================================================
-// Venturi Geometry Definition
-// ============================================================================
-
-/// Venturi throat geometry configuration
-///
-/// Defines the shape of a Venturi in 2D with:
-/// - Inlet section (constant width)
-/// - Converging section (linear taper)
-/// - Throat section (constant width)
-/// - Diverging section (linear expansion)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VenturiGeometry<T: Cfd2dScalar + Copy> {
-    /// Inlet width \[m]
-    pub w_inlet: T,
-    /// Throat width \[m]
-    pub w_throat: T,
-    /// Inlet section length \[m]
-    pub l_inlet: T,
-    /// Converging section length \[m]
-    pub l_converge: T,
-    /// Throat section length \[m]
-    pub l_throat: T,
-    /// Diverging (recovery) section length \[m]
-    pub l_diverge: T,
-    /// Channel height (constant) \[m]
-    pub height: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> VenturiGeometry<T> {
-    /// Create standard ISO 5167 Venturi with area ratio 0.5
-    ///
-    /// # Configuration
-    ///
-    /// - Inlet width: 10 mm
-    /// - Throat width: 7.07 mm (√2 ratio for 0.5 area)
-    /// - Standard converging length: ~0.1 × inlet width
-    /// - Standard throat length: ~0.2 × inlet width
-    /// - Standard diverging length: ~0.3-0.5 × inlet width
-    pub fn iso_5167_standard() -> Self {
-        Self {
-            w_inlet: from_f64::<T>(10e-3),
-            w_throat: from_f64::<T>(7.07e-3),
-            l_inlet: from_f64::<T>(10e-3),
-            l_converge: from_f64::<T>(1e-3),
-            l_throat: from_f64::<T>(2e-3),
-            l_diverge: from_f64::<T>(3e-3),
-            height: from_f64::<T>(1.0e-3),
-        }
-    }
-
-    /// Create custom Venturi geometry
-    pub fn new(
-        w_inlet: T,
-        w_throat: T,
-        l_inlet: T,
-        l_converge: T,
-        l_throat: T,
-        l_diverge: T,
-        height: T,
-    ) -> Self {
-        Self {
-            w_inlet,
-            w_throat,
-            l_inlet,
-            l_converge,
-            l_throat,
-            l_diverge,
-            height,
-        }
-    }
-
-    /// Explicitly opt into the built-in center-clustering recommendation for
-    /// `VenturiSolver2D::new_stretched`.
-    ///
-    /// This is a geometry-derived heuristic, not an implicit solver behavior.
-    /// Callers that want deterministic control should compute and pass their own
-    /// `beta` value instead of using this helper.
-    #[must_use]
-    pub fn recommended_center_clustering_beta(&self) -> T {
-        let four = from_f64::<T>(4.0);
-        let max_beta = from_f64::<T>(0.9);
-        let candidate = scalar::one::<T>() - four * self.w_throat / self.w_inlet;
-        <T as NumericElement>::min_scalar(
-            <T as NumericElement>::max_scalar(candidate, scalar::zero::<T>()),
-            max_beta,
-        )
-    }
-
-    /// Calculate area ratio (A_throat / A_inlet)
-    pub fn area_ratio(&self) -> T {
-        self.w_throat / self.w_inlet
-    }
-
-    /// Calculate total length
-    pub fn total_length(&self) -> T {
-        self.l_inlet + self.l_converge + self.l_throat + self.l_diverge
-    }
-
-    /// Get inlet cross-sectional area \[m²]
-    pub fn area_inlet(&self) -> T {
-        self.w_inlet * self.height
-    }
-
-    /// Get throat cross-sectional area \[m²]
-    pub fn area_throat(&self) -> T {
-        self.w_throat * self.height
-    }
-
-    /// Check if a point (x, y) is within the fluid domain
-    pub fn contains(&self, x: T, y: T) -> bool {
-        let _half_w_inlet = self.w_inlet / from_f64::<T>(2.0);
-        let _half_w_throat = self.w_throat / from_f64::<T>(2.0);
-
-        // X-ranges
-        let x_inlet_end = self.l_inlet;
-        let x_converge_end = x_inlet_end + self.l_converge;
-        let x_throat_end = x_converge_end + self.l_throat;
-        let x_diverge_end = x_throat_end + self.l_diverge;
-
-        if x < scalar::zero::<T>() || x > x_diverge_end {
-            return false;
-        }
-
-        // Calculate local width at x
-        let w_local = if x < x_inlet_end {
-            self.w_inlet
-        } else if x < x_converge_end {
-            // Linear convergence
-            let frac = (x - x_inlet_end) / self.l_converge;
-            self.w_inlet + frac * (self.w_throat - self.w_inlet)
-        } else if x < x_throat_end {
-            self.w_throat
-        } else if x < x_diverge_end {
-            // Linear divergence
-            let frac = (x - x_throat_end) / self.l_diverge;
-            self.w_throat + frac * (self.w_inlet - self.w_throat)
-        } else {
-            self.w_inlet
-        };
-
-        let half_w = w_local / from_f64::<T>(2.0);
-        eunomia::NumericElement::abs(y) <= half_w
-    }
-}
-
-// ============================================================================
-// Bernoulli-Based Analytical Solution
-// ============================================================================
-
-/// Analytical Venturi solution based on Bernoulli equation
-///
-/// Provides exact (frictionless) prediction of pressure distribution
-/// for validation against numerical solutions.
-pub struct BernoulliVenturi<T: Cfd2dScalar + Copy> {
-    geometry: VenturiGeometry<T>,
-    /// Inlet velocity \[m/s]
-    pub u_inlet: T,
-    /// Inlet pressure \[Pa]
-    pub p_inlet: T,
-    /// Fluid density [kg/m³]
-    pub rho: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> BernoulliVenturi<T> {
-    /// Create new Bernoulli solution
-    pub fn new(geometry: VenturiGeometry<T>, u_inlet: T, p_inlet: T, rho: T) -> Self {
-        Self {
-            geometry,
-            u_inlet,
-            p_inlet,
-            rho,
-        }
-    }
-
-    /// Calculate velocity at throat (from mass conservation)
-    ///
-    /// # Derivation
-    ///
-    /// Mass conservation: A₁u₁ = A₂u₂
-    /// ```text
-    /// u_throat = u_inlet × (A_inlet / A_throat)
-    ///          = u_inlet / area_ratio
-    /// ```
-    pub fn velocity_throat(&self) -> T {
-        self.u_inlet / self.geometry.area_ratio()
-    }
-
-    /// Calculate pressure at throat (from Bernoulli equation)
-    ///
-    /// # Derivation
-    ///
-    /// Bernoulli (frictionless): P + (1/2)ρu² = constant
-    /// ```text
-    /// P_throat = P_inlet + (1/2)ρ(u_inlet² - u_throat²)
-    ///          = P_inlet + (1/2)ρu_inlet²(1 - (A_inlet/A_throat)²)
-    /// ```
-    pub fn pressure_throat(&self) -> T {
-        let one_half = from_f64::<T>(0.5);
-        let u_throat = self.velocity_throat();
-        let dynamic_pressure_inlet = one_half * self.rho * self.u_inlet * self.u_inlet;
-        let dynamic_pressure_throat = one_half * self.rho * u_throat * u_throat;
-
-        self.p_inlet + (dynamic_pressure_inlet - dynamic_pressure_throat)
-    }
-
-    /// Calculate pressure coefficient at throat
-    ///
-    /// # Definition
-    ///
-    /// ```text
-    /// Cp = (P - P_inlet) / ((1/2)ρu_inlet²)
-    /// ```
-    ///
-    /// # Theoretical Value for Venturi
-    ///
-    /// ```text
-    /// Cp_ideal = 1 - (A_inlet / A_throat)²
-    ///          = 1 - 1/area_ratio²
-    /// ```
-    pub fn pressure_coefficient_throat(&self) -> T {
-        let ar = self.geometry.area_ratio();
-        let one = scalar::one::<T>();
-
-        // Cp_ideal = 1 - (1/area_ratio)²
-        one - (one / ar) * (one / ar)
-    }
-
-    /// Calculate pressure recovery in diffuser (outlet)
-    ///
-    /// For frictionless flow, outlet pressure should nearly equal inlet pressure.
-    /// For real (viscous) flow, there's some irreversible loss.
-    ///
-    /// # Pressure Recovery Ratio
-    ///
-    /// ```text
-    /// Cp_recovery = (P_outlet - P_inlet) / ((1/2)ρu_inlet²)
-    /// ```
-    ///
-    /// Ideal Venturi: Cp_recovery → 0 (full pressure recovery)
-    /// Real Venturi: Cp_recovery ~ -(0.05 to 0.15) (friction loss)
-    pub fn pressure_recovery_ideal(&self) -> T {
-        // For ideal flow with full recovery
-        scalar::zero::<T>()
-    }
-}
-
-// ============================================================================
-// Viscous Venturi Solution with Friction Loss
-// ============================================================================
-
-/// Venturi with viscous friction loss correction
-///
-/// Real Venturi flows have friction losses that reduce pressure recovery.
-/// This is modeled using a recovery coefficient C_r:
-///
-/// ```text
-/// P_outlet = P_inlet - ζ × (1/2)ρu_inlet²
-/// ```
-///
-/// where ζ is the loss coefficient (0.1-0.3 for typical Venturi).
-pub struct ViscousVenturi<T: Cfd2dScalar + Copy> {
-    bernoulli: BernoulliVenturi<T>,
-    /// Loss coefficient (typical: 0.1-0.3)
-    pub loss_coefficient: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> ViscousVenturi<T> {
-    /// Create Venturi with friction losses
-    ///
-    /// # Loss Coefficient
-    ///
-    /// Typical values for symmetric Venturi:
-    /// - Sharp convergence: ζ ≈ 0.05
-    /// - Smooth convergence: ζ ≈ 0.03
-    /// - Divergence (recovery): ζ ≈ 0.08-0.12
-    /// - Total: ζ ≈ 0.10-0.20
-    pub fn new(
-        geometry: VenturiGeometry<T>,
-        u_inlet: T,
-        p_inlet: T,
-        rho: T,
-        loss_coefficient: T,
-    ) -> Self {
-        let bernoulli = BernoulliVenturi::new(geometry, u_inlet, p_inlet, rho);
-        Self {
-            bernoulli,
-            loss_coefficient,
-        }
-    }
-
-    /// Calculate outlet pressure with friction loss
-    ///
-    /// # Formula
-    ///
-    /// ```text
-    /// P_outlet = P_inlet - ζ × (1/2)ρu_inlet²
-    /// ```
-    pub fn pressure_outlet_with_loss(&self) -> T {
-        let one_half = from_f64::<T>(0.5);
-        let dynamic_pressure_inlet =
-            one_half * self.bernoulli.rho * self.bernoulli.u_inlet * self.bernoulli.u_inlet;
-
-        self.bernoulli.p_inlet - self.loss_coefficient * dynamic_pressure_inlet
-    }
-
-    /// Calculate pressure recovery coefficient (real)
-    ///
-    /// # Definition
-    ///
-    /// ```text
-    /// Cp_recovery = (P_outlet - P_inlet) / ((1/2)ρu_inlet²)
-    ///             = -ζ
-    /// ```
-    pub fn pressure_recovery_coefficient(&self) -> T {
-        -self.loss_coefficient
-    }
-}
-
-// ============================================================================
-// Venturi Flow Solution (Numerical Result)
-// ============================================================================
-
-/// Solution to the Venturi flow problem
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct VenturiFlowSolution<T: Cfd2dScalar + Copy> {
-    /// Inlet velocity \[m/s]
-    pub u_inlet: T,
-    /// Inlet pressure \[Pa]
-    pub p_inlet: T,
-    /// Maximum throat velocity \[m/s] (centerline of parabolic profile)
-    pub u_throat: T,
-    /// Area-averaged throat velocity \[m/s].
-    ///
-    /// ## Theorem — Cross-Section Averaging
-    ///
-    /// For fully-developed laminar flow in a 2D channel, the velocity profile
-    /// is parabolic: u(y) = u_max · (1 − (2y/h)²), giving:
-    ///
-    /// u_mean = (1/h) ∫₀ʰ u(y) dy = (2/3) u_max
-    ///
-    /// The area-averaged velocity is the correct metric for comparison with the
-    /// 1D Hagen-Poiseuille / Bernoulli continuity prediction, which predicts:
-    ///
-    /// u_mean_throat = u_mean_inlet · (A_inlet / A_throat)
-    ///
-    /// **Reference**: White, F.M. (2011), *Fluid Mechanics* §6.3.
-    pub u_throat_mean: T,
-    /// Throat pressure \[Pa]
-    pub p_throat: T,
-    /// Outlet velocity \[m/s]
-    pub u_outlet: T,
-    /// Outlet pressure \[Pa]
-    pub p_outlet: T,
-    /// Pressure drop in throat \[Pa]
-    pub dp_throat: T,
-    /// Pressure recovery (outlet - inlet) \[Pa]
-    pub dp_recovery: T,
-    /// Pressure coefficient at throat
-    pub cp_throat: T,
-    /// Pressure recovery coefficient at outlet
-    pub cp_recovery: T,
-    /// Whether the iterative solver converged (always true for Bernoulli)
-    pub converged: bool,
-    /// SIMPLE iterations consumed (0 for the analytical Bernoulli path).
-    pub iterations: usize,
-    /// Final continuity residual reported by the iterative solver
-    /// (max of the RMS and mean-L1 per-cell mass imbalance; exactly zero
-    /// for the analytical Bernoulli path).
-    pub final_residual: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> VenturiFlowSolution<T> {
-    /// Create Venturi solution from Bernoulli
-    pub fn from_bernoulli(bernoulli: &BernoulliVenturi<T>, p_outlet: T) -> Self {
-        let u_throat = bernoulli.velocity_throat();
-        let p_throat = bernoulli.pressure_throat();
-        let cp_throat = bernoulli.pressure_coefficient_throat();
-
-        let one_half = from_f64::<T>(0.5);
-        let dynamic_pressure = one_half * bernoulli.rho * bernoulli.u_inlet * bernoulli.u_inlet;
-        let cp_recovery = (p_outlet - bernoulli.p_inlet)
-            / <T as NumericElement>::max_scalar(dynamic_pressure, from_f64::<T>(1.0));
-
-        Self {
-            u_inlet: bernoulli.u_inlet,
-            p_inlet: bernoulli.p_inlet,
-            u_throat,
-            u_throat_mean: u_throat, // Bernoulli is 1D: mean = max
-            p_throat,
-            u_outlet: bernoulli.u_inlet, // Mass conservation with constant height
-            p_outlet,
-            dp_throat: p_throat - bernoulli.p_inlet,
-            dp_recovery: p_outlet - bernoulli.p_inlet,
-            cp_throat,
-            cp_recovery,
-            converged: true,
-            iterations: 0,
-            final_residual: scalar::zero::<T>(),
-        }
-    }
-
-    /// Verify energy conservation (should account for dissipation)
-    ///
-    /// # Energy Balance
-    ///
-    /// For viscous flow:
-    /// ```text
-    /// P_inlet + (1/2)ρu_inlet² = P_outlet + (1/2)ρu_outlet² + ε_dissipated
-    /// ```
-    ///
-    /// Returns dissipated energy \[Pa]
-    pub fn energy_dissipation(&self, rho: T) -> T {
-        let one_half = from_f64::<T>(0.5);
-
-        let energy_inlet = self.p_inlet + one_half * rho * self.u_inlet * self.u_inlet;
-        let energy_outlet = self.p_outlet + one_half * rho * self.u_outlet * self.u_outlet;
-
-        <T as NumericElement>::max_scalar(energy_inlet - energy_outlet, scalar::zero::<T>())
-        // Dissipation is positive
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use eunomia::assert_relative_eq;
-
-    #[test]
-    fn test_venturi_geometry_iso() {
-        let geom = VenturiGeometry::<f64>::iso_5167_standard();
-
-        assert!(geom.w_inlet > geom.w_throat);
-        assert!(geom.area_ratio() < 1.0);
-        assert_relative_eq!(geom.area_ratio(), 0.707, epsilon = 0.01);
-    }
-
-    #[test]
-    fn test_bernoulli_venturi_mass_conservation() {
-        let geom = VenturiGeometry::<f64>::iso_5167_standard();
-        let bernoulli = BernoulliVenturi::new(geom.clone(), 1.0, 101325.0, 1000.0);
-
-        // Mass conservation: A1*u1 = A2*u2
-        let q_inlet = geom.area_inlet() * bernoulli.u_inlet;
-        let q_throat = geom.area_throat() * bernoulli.velocity_throat();
-
-        assert_relative_eq!(q_inlet, q_throat, epsilon = 1e-10);
-    }
-
-    #[test]
-    fn test_bernoulli_pressure_drop() {
-        let geom = VenturiGeometry::<f64>::iso_5167_standard();
-        let bernoulli = BernoulliVenturi::new(geom, 2.0, 101325.0, 1000.0);
-
-        let p_throat = bernoulli.pressure_throat();
-
-        // Pressure should decrease in throat
-        assert!(p_throat < bernoulli.p_inlet);
-    }
-
-    #[test]
-    fn test_viscous_venturi_recovery_loss() {
-        let geom = VenturiGeometry::<f64>::iso_5167_standard();
-        let viscous = ViscousVenturi::new(geom, 1.0, 101325.0, 1000.0, 0.15);
-
-        let p_outlet = viscous.pressure_outlet_with_loss();
-
-        // Outlet pressure should be less than inlet due to loss
-        assert!(p_outlet < viscous.bernoulli.p_inlet);
-    }
-}
